ThemeProcess Tests fixed
